requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
nltk>=3.6
scikit-learn>=1.0.0
numpy>=1.20.0
PyPDF2>=2.0.0
python-docx>=0.8.0
bleach>=4.0.0
setuptools>=60.0.0
wheel>=0.37.0